frontend/requirements.txt
Flask==2.2.5
python-dotenv==0.21.0
requests==2.31.0
flask-jwt-extended==4.3.0